+++ apps/server/src/db/reindex-search.ts
@@ -1,6 +1,6 @@
 import { loadEnv } from '../env.ts'
 import { createDb } from './client.ts'
-import { rebuildSearchIndex } from '../services/search.ts'
+import { rebuildSearchIndex } from './repositories/search.ts'
 
 const env = loadEnv()
 const db = createDb(env.database, { ftsTokenizer: env.search.ftsTokenizer })